Reference

API

NMRlab Basics

NMRlab.FourierTransform — Method

function FourierTransform(SI::Vector,dims::Vector; fftshift=true)

Fourier transform processor for data sets of size SI. dims is a Vector of the dimensions along which a Fourier transform will be computed. The corresponding coordinates are automatically replaced by frequencies, based on the Nyqvist theorem. The zero frequency appears in the centre of the spectrum.

NMRlab.MedianBaselineCorrect — Method

function (mb::MedianBaselineCorrect)(s::SpectData)

subtract baseline for the real part of s by the algorithm of M. S. Friedrichs, Journal of Biomolecular NMR, 5 (1995) 147 153.

NMRlab.NMRProcessor — Type

abstract type NMRProcessor <: Function end

Abstract data type for NMR processing routines. New processing tools should be declared as subtypes of NMRProcessor.

NMRlab.NMRProcessor — Method

(m::NMRProcessor)(A::AbstractArray)

here is some fallback behaviour. NMR processors are defined for SpectData objects. However, they can also be applied to any AbstractArray, by promoting it to a SpectData, and then extracting the data part.

NMRlab.NMRProcessor1D — Type

abstract type NMRProcessor1D <: NMRProcessor

Data type for processing tools that apply to a single dimension, i.e., that are inherently 1D. They need to be defined as functors that act on a AbstractVector{T}. They must contain a field .dim that indicates which dimension in a multidimensional array they should be applied to.

NMRlab.SpectData — Type

struct SpectData{T,N} <: AbstractArray{T,N}

Basic structure for spectral data.

NMRlab.Chain — Method

Chain(fs::Vararg{NMRProcessor})

returns a chain of processing tools, which will be applied in order (the first in the argument list is applied first)

NMRlab.conv — Method

function conv(X::AbstractArray{T1,N}, y::AbstractVector{T2},dim::Integer) where {N>1,T1,T2}

computes the convolution of the array X with the vector y along the dimension dim. The ends of X are zero-padded such that the result is guaranteed to have the same size as X. NMRlab.conv() uses a direct algorithm for the convolution, not fft. It is therefore efficient when the length of y is much less than the corresponding dimension of X. If this is not the case and performance is critical, a different algorithm should be used.

NMRlab.coords — Method

function coords(S::SpectData)
function coords(S::SpectData,k::Integer)

returns a tuple with the coordinates of S, analogous to axes(S). The second form returns the coordinate of the k-th dimension.

NMRlab.extrema — Method

function extrema(X::AbstractArray{T,N}, dim::Integer)

returns an array of all extremal values of X along the dimension dim.

NMRlab.find_spdta — Method

A = find_spdta(As) returns the first SpectData among the arguments.

NMRlab.load — Method

function load(path::String,vendor::Symbol)

Load a data set located at path, and return a dictionary with the acquisition parameters as well as a SpectData object with the actual (time domain) data. vendor designates the origin data format. Currently implemented are

:Bruker: the path points to a directory with a Bruker NMR data set.

:JEOL: the path points to a JEOL .jdf file

NMRlab low-level File IO

NMRlab.FileIO — Module

module FileIO

contains low-level code to import vendor-specific NMR data files.

NMRlab.FileIO.readBrukerFID — Method

Bruker.readBrukerFID(s::String; format=Float64)

reads a Bruker fid or ser file and returns the data contained in it as a complex array. For files generated by TopSpin 4.0 and higher, the default format (Float64) will work. For TopSpin 2.0, use format=Int32.

NMRlab.FileIO.readBrukerParameterFile — Method

readBrukerParameterFile(s::String)

reads a Bruker JCAMP-DX parameter file, and returns a dict with the values. Arrays such as D0 D1 D2 etc are returned as a Julia array. Note that the Bruker parameter D0 appears as param["D"][1] due to Julia using 1-based arrays.

NMRlab.FileIO.readJEOL — Method

function readJEOL(s::IOStream) -> (header,params,data)

reads a JEOL .jdf file from the stream s. header contains a dictionary with the file header values, params is a dictionary of parameters. Each parameter contains a tuple (scaler::Int,units,value).

data is a raw data vector, which needs to be reshaped into the correct form depending on the parameters. This is handled by a separate function.

NMRlab.FileIO.reshapeJEOL — Method

function reshapeJEOL(header,params,data<:AbstractArray)

uses the header and parameter data to reshape the data. Returns an Array with the correctly shaped data.

Not implemented yet!

This function is not yet implemented. It will rely on the new data type, and will return a valid SpectData array, including the correct axes and coordinate information.

Spin Simulation

NMRlab.SpinSim.PeakSpect — Method

function PeakSpect(p,i,r;lw=0.0001)

compute the the spectrum with Lorentzian peaks at locations given in array p and (complex) amplitudes in array i, over the range r, which can be either an array or a range. lwis the line width. AData1D` object is returned.

NMRlab.SpinSim.Spectrum — Method

    function Spectrum(ρ,H,Ψ;tol=1e-3)

compute the frequencies and intensities of the spectrum given by the the initial density operator ρ, the Hamiltonian H, and the observation operator Ψ. The results are returned as a tuple (freq,int) of two one-dimensional arrays. tol is a cutoff; transitions with absolute amplitudes less than this value are suppressed.

NMRlab.SpinSim.expm — Method

function expm(A::SparseMatrixCSC)

compute the matrix exponential of A, using the dividing and squaring algorithm, as described by I. Kuprov, "Spin", Springer, Zürich, 2024, pp 173ff.

GISSMO

NMRlab.GISSMO — Module

module GISSMO

Interface routines with GISSMO database.

Hesam Dashti, Jonathan R. Wedell, William M. Westler, Marco Tonelli, David Aceti, Gaya K. Amarasinghe, John L. Markley, and Hamid R. Eghbalnia, Applications of Parametrized NMR Spin Systems of Small Molecules, Anal. Chem., 2018, 90 (18), pp 10646–10649, DOI: 10.1021/acs.analchem.8b02660 Hesam Dashti, William M. Westler, Marco Tonelli, Jonathan R. Wedell, John L. Markley, and Hamid R. Eghbalnia, Spin System Modeling of Nuclear Magnetic Resonance Spectra for Applications in Metabolomics and Small Molecule Screening, Analytical Chemistry, 2017, 89 (22), pp 12201–12208, doi: 10.1021/acs.analchem.7b02884

NMRlab.GISSMO.Hamiltonian — Method

function Hamiltonian(fn::String;freq=600.0,ctr=4.8)
function Hamiltonian(system::XMLElement;freq=600.0,ctr=4.8)

computes the Hamiltonian of an entry in the GISSMO database and returns it in the natural basis as a sparse matrix. fn is the GISSMO reference name of the compound. Keyword parameters are used to indicate spectrometer base frequency (in MHz) and the spectral zero point (carrier position) in ppm. maxSpin represents the maximum size of the spin system that will be simulated. If the spin system size exceeds maxSpin, the function returns (-1,nothing).

NMRlab.GISSMO.SpinMatrix — Method

function SpinMatrix(id::String)

returns the spin matrix, with chemical shifts (on the diagonal) and J-couplings (on the off-diagonal elements) for the GISSMO entry specified by the identifier id.

NMRlab.GISSMO.search — Method

function search(term::String)

searches the online GISSMO database for term and returns a JSON object with the search result.

Index

NMRlab.FileIO

NMRlab.GISSMO

NMRlab.FourierTransform

NMRlab.MedianBaselineCorrect

NMRlab.NMRProcessor

NMRlab.NMRProcessor

NMRlab.NMRProcessor1D

NMRlab.SpectData

NMRlab.Chain

NMRlab.FileIO.readBrukerFID

NMRlab.FileIO.readBrukerParameterFile

NMRlab.FileIO.readJEOL

NMRlab.FileIO.reshapeJEOL

NMRlab.GISSMO.Hamiltonian

NMRlab.GISSMO.SpinMatrix

NMRlab.GISSMO.search

NMRlab.SpinSim.PeakSpect

NMRlab.SpinSim.Spectrum

NMRlab.SpinSim.expm

NMRlab.conv

NMRlab.coords

NMRlab.extrema

NMRlab.find_spdta

NMRlab.load